+++ .gitignore
@@ -2,6 +2,7 @@
 
 # Hydra
 outputs/
+multirun/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
