+++ examples/simple_run.py
@@ -1,4 +1,4 @@
-from benchmarl.algorithms.mappo import MappoConfig
+from benchmarl.algorithms import MaddpgConfig, MappoConfig, MasacConfig, QmixConfig
 from benchmarl.benchmark import Benchmark
 from benchmarl.environments import VmasTask
 from benchmarl.experiment import ExperimentConfig
@@ -7,10 +7,17 @@ from benchmarl.models.mlp import MlpConfig
 
 if __name__ == "__main__":
 
+    experiment_config = ExperimentConfig.get_from_yaml()
     tasks = [VmasTask.BALANCE.get_from_yaml()]
-    algorithm_configs = [MappoConfig.get_from_yaml()]
+    algorithm_configs = [
+        MappoConfig.get_from_yaml(),
+        MaddpgConfig.get_from_yaml(),
+        QmixConfig.get_from_yaml(),
+        MasacConfig.get_from_yaml(),
+    ]
     seeds = {0}
 
+    # Model still need to be refactored for hydra loading
     model_config = SequenceModelConfig(
         model_configs=[
             MlpConfig(num_cells=[64, 64]),
@@ -18,7 +25,6 @@ if __name__ == "__main__":
         ],
         intermediate_sizes=[128],
     )
-    experiment_config = ExperimentConfig.get_from_yaml()
 
     benchmark = Benchmark(
         algorithm_configs=algorithm_configs,
